+++ metricbeat/module/zookeeper/_meta/docs.asciidoc
@@ -1,5 +1,5 @@
 The ZooKeeper module fetches statistics from the ZooKeeper service. The default
-metricset is `mntr`. You can also activate `server` metricset.
+metricset is `mntr` and `server`.
 
 [float]
 === Compatibility
