+++ metricbeat/module/zookeeper/fields.go
@@ -32,5 +32,5 @@ func init() {
 // AssetZookeeper returns asset data.
 // This is the base64 encoded gzipped contents of ../metricbeat/module/zookeeper.
 func AssetZookeeper() string {
-	return "eJy0mEuP4zYMx+/5FMRe9rT5ADkUKLo9FMXuAtOe9uJlZCYWRhZdic5jPn0hvx9yJq/xaWBH/P9IkSI1X+CVzht4Y34lKsitAESLoQ18+sn8d/Xu0wogJa+cLkSz3cBvKwCA7jvkJE4rD4qNISWUwvYMkhHsuHRfDImEH7HVwk7bPSjOc7SpX68AfMZOEsV2p/cb2KHxtAJwZAg9bWCPK4CdJpP6TaX6BSzmNCYOj5yL8HPHZdG8iSCH51e38hcotoLa+gq29cJRwa5xovOxWz5kD8+QbciXW+kXxfAuIFaYwcAVhNMw1+tGphrk9eDlOL7tM/Vl6E/GXsJfo4+tX690PrJLJ98ueDfOn9b2OqqMReH4pHMUSlIUTLx+i2MYtvvbGH7vbUMwC7wbcAW1OJNBIavOazxMBe8mOZDDPbWWYUtyJLJA1lO+NVQFyYO2kGtjtCfFgywc01GRUU4OjU8Ul1aehPi9zLfkQog6AXiznNICxo6N4SM5/3T9zjL4EKOmDFTpHFmpIhUnyvGU7LShpFVi99QIfcOTzsscbE+qDUGv5gEr8hR27CroPtsKx4r8QizbhMvx9GTWNuGuyqwOQ9tnYWh7O4Yt8wSNPlBoHJZUMP78LBvYBuHBTkmGAui6jDNnqGjisFyQ/dCs+76cbUG7LY8rM41L8YI21XafOPqvJC/PD+1ABFqROqyWKA3R3lKL2fc5ZUov5OLcBapXEr92pEgfaNqMHmfu42dJjuxeW0loJS+DefqAHV+GCnILQFTvrj9b9fytbaxDZT3spELnzmHHx0XUNbbuMI/TenIHcokXlCcOHy9sKJw148JomRZIzlZRmnxcW6sFBt2tG/SOGVlACM12FBDQHgxhulQUB3Jec/ysRqNx6kOBktUh14rW8dVXj3XNckCbwrbUJgUvrq742q048xFFZR90QFa2yQ8Obk8CXOeBYYXm2mOyGnw+iLIeqmZj/rt845K57fbxuV50x4XDu0N1UakvGZ8fumU82s7/GLRs8oJbo33Wc88CM5tsoqrTwC15MbQ3vxlc9GLmyfQ+wLslD2A84D6k+g1PnaIjVJdjN1KeTYQ3Kk/mwGoYyNDDNkz4UZbmZXzS5/TubvFPXQkzE930+Vjl/8uCZnBFqIr9vVHsLqUfC+tb+w+NSi/N4sH8xu8V2d0j0EurESxcUc2X+t491aynMerN1T3tvfT+2f7TqeuKrrQ2trCVTOfzTi+68HEk+TXMB82x0cs3Z3KYfUofz+/TzNv7A8dO77VFs+hJvB5nzvxo7MABTRlxSxxaj9XJD399XcSJVS3cdE7V1TuQ8y1MPYeFwFLBKltkiH+9heHPYOHKSPwfAAD///RUANo="
+	return "eJy0mM2O4zYMx+95CmIve5o8QA4Fim4PRbG7wLSnvXgYmYmFyKIr0fmYpy9kW/6Kncnkw6eBPeL/R4oUqbzAjk4reGfeERXkFgCixdAKvvxi/rt692UBkJJXThei2a7gtwUAQPsdchKnlQfFxpASSmF9AskINly6F0Mi4Z/YamGn7RYU5zna1C8XAD5jJ4liu9HbFWzQeFoAODKEnlawxQXARpNJ/apSfQGLOQ2JwyOnIvy747Jo3kwgh+etXfkGiq2gtr6CjV44Ktg1TrQ+tsv77OHps/X5civdoim8C4gVZjBwBeE4zPW6gakGedl7OYxvfMa+9P3J2Ev4a/Ax+rWj04FdOvp2wbth/kTby0llLArHR52jUJKiYOL1+zSGYbv9HMPvnW0IZoE3Pa6gNs1kUMiq0xL3Y8GbSfbkcEvRMqxJDkQWyHrK14aqIHnQFnJtjPakuJeFQzoqMsrJofGJ4tLKgxB/lPmaXAhRKwDvllOawdiwMXwg5x+u31oGH2LUlIEqnSMrVaSmiXI8JhttKIlK7B4aoe941HmZg+1ItSHo1DxgRZ7Chl0F3WVb4ViRn4llTLgcjw9mjQl3VWa1GNo+CkPbz2PYMk/Q6D2FxmFJBeOPz7KebRDu7ZRkKICuzThzgopmGpYLsk/Nuh/z2Ra0Y3lcmWlcihe0qbbbxNF/JXl5fGh7IhBF6rBaojREe00Rs+tzypReyE1zF6h2JH7pSJHe07gZ3c/cxc+SHNjtoiREyctgnp6w4/NQQW4GiOrd9SerHr+1jXWorIedVOjcKez4sIjaxtYe5tO0ntyeXOIF5YHDxysbCmfNsDAi0wzJySpKk+e1tVqg193aQe+QkQWE0GwHAQHtwRCmc0WxJ+c1T5/VaDSOfShQsjrkWtFyevXVY12zHNCmsC61ScGLqyu+dmua+YCisicdkJVt8r2D25MA13lgWKG59pisBp8nUdZD1dmY/yHfsGQ+d/v4Wi+64cLh3b66qNSXjK933TLubed/9Fo2ecG10T7ruM8CczbZTKqOAzfnRd/e+c3gohdnnozvA7yZ8wCGA+5dqt/x2Co6QnU5dgPls4nwk8qjObAaBjL0sA4T/iRL83J60uf05m7xT10JZyba6fO+yv+XBU3vilAV+0ej2E1KP2fWR/t3jUqvzeLe/MYfFdnNI9Br1AgWrqjmS33vlmrW4xh15uqe9lF6/4o/OrVd0ZXWTi2Mkun5vNOJznwcSH4L80FzbHTyzZkcZp/ST+f3+/HM3atr56fTW23RwB5NOSEvDq3H6oSGv77N9ID7Kqsn4SNAPSMFp6lglc38ajL+cq3un2HllR7/HwAA//8U/OsQ"
 }
