+++ metricbeat/module/zookeeper/server/_meta/docs.asciidoc
@@ -11,6 +11,6 @@
 * *sent*: Requests sent by the server
 * *version.id*: Zookeeper version running
 * *version.date*: Date of the Zookeeper release in use
-* *xid.original*: Original value of the Zookeeper transaction ID
-* *xid.count*: Total transactions of the leader in epoch
-* *xid.epoch*: Epoch value of the Zookeeper transaction ID
+* *zxid*: Original value of the Zookeeper transaction ID
+* *count*: Total transactions of the leader in epoch
+* *epoch*: Epoch value of the Zookeeper transaction ID
